// src/types/inputType.tsx
import { ReactElement } from "react";
export interface InputProps {
  color?: any;
  fontSize?: any;
  fontWeight?: any;
  lineHeight?: any;
  padding?: any;
  placeholder?: any;
  margin?: any;
  width?: any;
  border?: any;
  borderRadius?: any;
  className?: any;
  inputIcon?: ReactElement;
  position?: any;
  checkMultiLine?: boolean;
  maxRows?: number;
  alignItems?: string;
  height?: string;
  onChange?: any;
  value?: any;
  type?: any;
  ref?: any;
  name?: any;
}